requirements.txt
pandas
requests
dnspython
beautifulsoup4
tldextract
urllib3